Moved model and model meta loading utilities into overview.py
